Don't include kis_image_signal_router.h in headers
The final goal is to split KisImage into multiple classes to avoid
full recompilation on every change in kis_image.h. KisImageSignalRouter
will contain all the image's signals.
