+++ libs/image/kis_image_signal_router.h
@@ -20,119 +20,11 @@
 #define __KIS_IMAGE_SIGNAL_ROUTER_H
 
 #include <QObject>
-#include <QVector>
-#include <QRectF>
-#include "kritaimage_export.h"
-#include "kis_types.h"
-
+#include "KisImageSignals.h"
 
 class KoColorSpace;
 class KoColorProfile;
 
-enum KisImageSignalTypeEnum {
-    LayersChangedSignal,
-    ModifiedSignal,
-    SizeChangedSignal,
-    ProfileChangedSignal,
-    ColorSpaceChangedSignal,
-    ResolutionChangedSignal,
-    NodeReselectionRequestSignal
-};
-
-/**
- * A special signal which handles stillPoint capabilities of the image
- *
- * \see KisImage::sigSizeChanged()
- */
-struct ComplexSizeChangedSignal {
-    ComplexSizeChangedSignal() {}
-    ComplexSizeChangedSignal(QPointF _oldStillPoint, QPointF _newStillPoint)
-        : oldStillPoint(_oldStillPoint),
-          newStillPoint(_newStillPoint)
-    {
-    }
-
-    /**
-     * A helper method calculating the still points from image areas
-     * we process. It works as if the source image was "cropped" by \p
-     * portionOfOldImage, and this portion formed the new image of size
-     * \p transformedIntoImageOfSize.
-     *
-     * Note, that \p portionOfTheImage may be equal to the image bounds().
-     */
-    ComplexSizeChangedSignal(const QRect &portionOfOldImage, const QSize &transformedIntoImageOfSize)
-    {
-        oldStillPoint = QRectF(portionOfOldImage).center();
-        newStillPoint = QRectF(QPointF(), QSizeF(transformedIntoImageOfSize)).center();
-    }
-
-    ComplexSizeChangedSignal inverted() const {
-        return ComplexSizeChangedSignal(newStillPoint, oldStillPoint);
-    }
-
-    QPointF oldStillPoint;
-    QPointF newStillPoint;
-};
-
-/**
- * A special signal which handles emitting signals for node reselection
- *
- * \see KisImage::sigRequestNodeReselection()
- */
-struct ComplexNodeReselectionSignal {
-    ComplexNodeReselectionSignal() {}
-    ComplexNodeReselectionSignal(KisNodeSP _newActiveNode, KisNodeList _newSelectedNodes,
-                                 KisNodeSP _oldActiveNode = KisNodeSP(), KisNodeList _oldSelectedNodes = KisNodeList())
-        : newActiveNode(_newActiveNode),
-          newSelectedNodes(_newSelectedNodes),
-          oldActiveNode(_oldActiveNode),
-          oldSelectedNodes(_oldSelectedNodes)
-    {
-    }
-
-    ComplexNodeReselectionSignal inverted() const {
-        return ComplexNodeReselectionSignal(oldActiveNode, oldSelectedNodes, newActiveNode, newSelectedNodes);
-    }
-
-    KisNodeSP newActiveNode;
-    KisNodeList newSelectedNodes;
-    KisNodeSP oldActiveNode;
-    KisNodeList oldSelectedNodes;
-};
-
-struct KisImageSignalType {
-    KisImageSignalType() {}
-    KisImageSignalType(KisImageSignalTypeEnum _id)
-    : id(_id)
-    {
-    }
-
-    KisImageSignalType(ComplexSizeChangedSignal signal)
-        : id(SizeChangedSignal),
-          sizeChangedSignal(signal)
-    {
-    }
-
-    KisImageSignalType(ComplexNodeReselectionSignal signal)
-        : id(NodeReselectionRequestSignal),
-          nodeReselectionSignal(signal)
-    {
-    }
-
-    KisImageSignalType inverted() const {
-        KisImageSignalType t;
-        t.id = id;
-        t.sizeChangedSignal = sizeChangedSignal.inverted();
-        t.nodeReselectionSignal = nodeReselectionSignal.inverted();
-        return t;
-    }
-
-    KisImageSignalTypeEnum id;
-    ComplexSizeChangedSignal sizeChangedSignal;
-    ComplexNodeReselectionSignal nodeReselectionSignal;
-};
-
-typedef QVector<KisImageSignalType> KisImageSignalVector;
 
 class KRITAIMAGE_EXPORT KisImageSignalRouter : public QObject
 {
